Move implementations to entities
